requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
groq==0.4.2